requirements.txt
pandas
snowflake-connector-python
pyyaml
pytest
